Split presentation frame state
